Move detector into htmresearch, s/NAB/nab_experiments
