+++ projects/nab_experiments/results/.gitignore
@@ -1,2 +1,3 @@
 *
+!.gitignore
 !final_results.json
